refactor: renames metrics module for generality
